requirements.txt
streamlit
openai
streamlit-chat
openai
watchdog
PyMuPDF
python-docx
langchain
faiss-cpu
langchain-community
tiktoken